Struct random_oracle::RandomOracle

#[repr(transparent)]
pub struct RandomOracle(_);

State of the random oracle, used to incrementally build up the output.

Implementations

impl RandomOracle

pub fn empty() -> Self

Start with the initial empty state of the oracle.

pub fn domain<B: AsRef<[u8]>>(data: B) -> Self

Start with the initial domain string.

pub fn split(&self) -> Self

Duplicate the random oracle, creating a fresh copy of it. Further updates are independent.

pub fn add<B: Serial>(&mut self, data: &B)

Append the input to the state of the oracle.

pub fn add_bytes<B: AsRef<[u8]>>(&mut self, data: B)

pub fn append_message<S: Serial, B: AsRef<[u8]>>(
&mut self,
label: B,
message: &S
)

Append the input to the state of the oracle, using label as domain separation.

pub fn extend_from<'a, I, S: 'a, B: AsRef<[u8]>>(&mut self, label: B, iter: I) where
S: Serial,
I: IntoIterator<Item = &'a S>,

Append all items from an iterator to the random oracle. Equivalent to repeatedly calling append in sequence. Returns the new state of the random oracle, consuming the initial state.

pub fn result_to_scalar<C: Curve>(self) -> C::Scalar

Try to convert the computed result into a field element. This interprets the output of the random oracle as a big-endian integer and reduces is mod field order.

pub fn get_challenge(self) -> Challenge

Get a challenge from the current state, consuming the state.

pub fn challenge_scalar<C: Curve, B: AsRef<[u8]>>(
&mut self,
label: B
) -> C::Scalar

Get a challenge in the form of a Scalar, using label as domain separation.